# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "ProductsController_getAllProducts",
        "summary": "Get all products for current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of products retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["products"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ProductsController_createProduct",
        "summary": "Create a new product",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["products"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_getProduct",
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": ["products"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductsController_updateProduct",
        "summary": "Update an existing product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": ["products"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_deleteProduct",
        "summary": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": ["products"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products/{id}/metadata": {
      "post": {
        "operationId": "ProductsController_addMetadata",
        "summary": "Add metadata to a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMetadataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Metadata added successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": ["products"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/settings/business": {
      "get": {
        "operationId": "SettingsController_getBusinessInfo",
        "summary": "Get business information",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Business information retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Business information not found"
          }
        },
        "tags": ["settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateBusinessInfo",
        "summary": "Update business information",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Business information updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/settings/payment-methods": {
      "get": {
        "operationId": "SettingsController_getPaymentMethods",
        "summary": "Get all payment methods",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Payment methods retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "SettingsController_createPaymentMethod",
        "summary": "Create a new payment method",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment method created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/settings/payment-methods/{id}": {
      "patch": {
        "operationId": "SettingsController_updatePaymentMethod",
        "summary": "Update a payment method",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment method updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Payment method not found"
          }
        },
        "tags": ["settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SettingsController_deletePaymentMethod",
        "summary": "Delete a payment method",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment method deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Payment method not found"
          }
        },
        "tags": ["settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "iPhone 15 Pro"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Latest iPhone model with advanced features"
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 999.99
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "XAF",
            "default": "XAF"
          },
          "category": {
            "type": "string",
            "description": "Product category",
            "example": "Electronics"
          },
          "images": {
            "description": "Product images URLs",
            "example": [
              "https://example.com/image1.jpg",
              "https://example.com/image2.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["name"]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "iPhone 15 Pro"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Latest iPhone model with advanced features"
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 999.99
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "XAF",
            "default": "XAF"
          },
          "category": {
            "type": "string",
            "description": "Product category",
            "example": "Electronics"
          },
          "images": {
            "description": "Product images URLs",
            "example": [
              "https://example.com/image1.jpg",
              "https://example.com/image2.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AddMetadataDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Metadata key",
            "example": "size"
          },
          "value": {
            "type": "string",
            "description": "Metadata value",
            "example": "Large"
          },
          "isVisible": {
            "type": "boolean",
            "description": "Whether the metadata is visible to customers",
            "example": true
          }
        },
        "required": ["key", "value", "isVisible"]
      },
      "UpdateBusinessInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Business name",
            "example": "My Business"
          },
          "description": {
            "type": "string",
            "description": "Business description",
            "example": "We provide quality products and services"
          },
          "address": {
            "type": "string",
            "description": "Business address",
            "example": "123 Main Street"
          },
          "city": {
            "type": "string",
            "description": "Business city",
            "example": "Douala"
          },
          "country": {
            "type": "string",
            "description": "Business country",
            "example": "Cameroon"
          },
          "website": {
            "type": "string",
            "description": "Business website",
            "example": "https://mybusiness.com"
          },
          "openingHours": {
            "type": "object",
            "description": "Opening hours in JSON format",
            "example": {
              "monday": "09:00-18:00",
              "tuesday": "09:00-18:00",
              "wednesday": "09:00-18:00",
              "thursday": "09:00-18:00",
              "friday": "09:00-18:00",
              "saturday": "09:00-14:00",
              "sunday": "closed"
            }
          },
          "phoneNumbers": {
            "description": "Additional phone numbers",
            "example": ["+237699999999", "+237600000000"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreatePaymentMethodDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Payment method type",
            "enum": ["CASH", "MOBILE_MONEY"],
            "example": "MOBILE_MONEY"
          },
          "mobileMoneyNumber": {
            "type": "string",
            "description": "Mobile money number (required if type is MOBILE_MONEY)",
            "example": "+237699999999"
          },
          "mobileMoneyName": {
            "type": "string",
            "description": "Mobile money provider name (e.g., MTN, Orange)",
            "example": "MTN"
          },
          "requiresProof": {
            "type": "boolean",
            "description": "Whether proof of transfer is required",
            "example": true
          }
        },
        "required": ["type"]
      },
      "UpdatePaymentMethodDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Payment method type",
            "enum": ["CASH", "MOBILE_MONEY"],
            "example": "MOBILE_MONEY"
          },
          "mobileMoneyNumber": {
            "type": "string",
            "description": "Mobile money number (required if type is MOBILE_MONEY)",
            "example": "+237699999999"
          },
          "mobileMoneyName": {
            "type": "string",
            "description": "Mobile money provider name (e.g., MTN, Orange)",
            "example": "MTN"
          },
          "requiresProof": {
            "type": "boolean",
            "description": "Whether proof of transfer is required",
            "example": true
          }
        }
      },
...
    }
  }
}